+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.1.68
 
-- in progress
+- fixed bug in Coco.Collection
+- added babel/polyfill requirement for IE support
 
 ## 0.1.67